Adding one more level of abstraction: TrackingClient that handles all the shared concrete implementations.
